requirements.txt
streamlit
pandas
Pillow
plotly
psycopg2-binary
sqlalchemy
bcrypt
SQLAlchemy

